Move types to DataModel.jl
